requirements.txt
discord.py
os
python-dotenv
requests
random